requirements.txt
cffi==1.17.1
cryptography==41.0.3
exceptiongroup==1.2.2
fire==0.1.3
iniconfig==2.1.0
packaging==25.0
pickleDB==0.9.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
PyYAML==6.0.2
six==1.17.0
tomli==2.2.1
